[sandbox] Reuse TagRange and Payload struct for the CppHeapPointerTable
With https://crrev.com/c/6055043, the CppHeapPointerTable and the
ExternalPointerTable now use the same type-tagging scheme. With this CL,
the CppHeapPointerTable also now reuses the generic TagRange struct
instead of defining its own CppHeapPointerTagRange and both pointer
tables share a common Payload struct.

Bug: 381444303
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6066319
Commit-Queue: Samuel Groß <saelo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#104355}

+++ include/v8-internal.h
@@ -486,8 +486,8 @@ struct TagRange {
     // Need to perform the math with uint32_t. Otherwise, the uint16_ts would
     // be promoted to (signed) int, allowing the compiler to (wrongly) assume
     // that an underflow cannot happen as that would be undefined behavior.
-    return static_cast<uint32_t>(tag) - first <=
-           static_cast<uint32_t>(last) - first;
+    return static_cast<uint32_t>(tag) - static_cast<uint32_t>(first) <=
+           static_cast<uint32_t>(last) - static_cast<uint32_t>(first);
   }
 
   constexpr bool Contains(TagRange tag_range) const {
